Refactoring outdated files and file structure
