--- src/game/card/rarities.json
@@ -1,30 +1,0 @@
-[
-    {
-      "color": "#ffffff",
-      "id": "common",
-      "name": "Common",
-      "value": 1,
-      "weight": 56
-    },
-    {
-      "color": "#00ff00",
-      "id": "uncommon",
-      "name": "Uncommon",
-      "value": 2,
-      "weight": 33
-    },
-    {
-      "color": "#0000ff",
-      "id": "rare",
-      "name": "Rare",
-      "value": 3,
-      "weight": 7
-    },
-    {
-      "color": "#0000ff",
-      "id": "rare holo",
-      "name": "Rare Holo",
-      "value": 4,
-      "weight": 4
-    }
-]
